rag-pinecone-docker/requirements.txt
langchain==0.1.16
langchain-core
langchain-community
streamlit
pinecone-client
pypdf
ollama